Revert "Sw12 155 admin school 2 edit pupil"
