--- foodplanner/lib/pages/admin_administration/EditClasses.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/material.dart';
-
-class EditClasses extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        toolbarHeight: 200,
-        centerTitle: true,
-        title: Padding(
-          padding: const EdgeInsets.only(top: 25),
-          child: Text(
-            'Redigere klasse',
-            style: TextStyle(fontSize: 36),
-            textAlign: TextAlign.center,
-          ),
-        ),
-      ),
-      backgroundColor: Colors.white,
-    );
-  }
-}
